refactor(playback profile): move files to more relevant folders
