+++ utils/playbackProfiles/helpers/fmp4VideoFormats.ts
@@ -2,8 +2,8 @@ import { hasH264Support, hasHevcSupport } from './mp4VideoFormats';
 import { browserDetector } from '~/plugins/browserDetection';
 
 /**
- * @param {HTMLVideoElement} videoTestElement s
- * @returns {string[]} List of ...
+ * @param {HTMLVideoElement} videoTestElement A HTML video element for testing codecs
+ * @returns {string[]} List of supported fmp4 video codecs
  */
 export function getSupportedFmp4VideoCodecs(
   videoTestElement: HTMLVideoElement
