+++ utils/playbackProfiles/helpers/tsVideoFormats.ts
@@ -1,8 +1,8 @@
 import { hasH264Support } from './mp4VideoFormats';
 
 /**
- * @param {HTMLVideoElement} videoTestElement s
- * @returns {string[]} List of ...
+ * @param {HTMLVideoElement} videoTestElement A HTML video element for testing codecs
+ * @returns {string[]} List of supported ts video codecs
  */
 export function getSupportedTsVideoCodecs(
   videoTestElement: HTMLVideoElement
